Update README.md
Firstly we declared our variables as integer values, after realizing that integer variables can't contain decimal values , we decided to make use of double variables
Secondly we prompted end users to provide the values for the base and the height.
Thirdly we added the area calculation , which is base multiplied by the height (base * height). 
The program displayed the end results.